backend/doc_intelligence/requirements.txt
openai
docling
camelot-py[cv]
pandas
python-dotenv
regex
tiktoken
jsonschema
Pillow
